requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
dask==2022.12.1
docker-pycreds==0.4.0
dpkt==1.9.8
duckdb==0.6.1
einops==0.6.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.51.1
hydra-core==1.3.0
idna==3.4
importlib-metadata==5.1.0
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.3.0
locket==1.0.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
packaging==21.3
pandas==1.5.2
pandasql==0.7.3
partd==1.3.0
pathtools==0.1.2
Pillow==9.3.0
promise==2.3
protobuf==3.20.1
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyshark==0.5.3
python-dateutil==2.8.2
pytorch-lightning==1.8.3.post1
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scapy==2.4.5
scikit-learn==1.2.0
scipy==1.9.3
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
SQLAlchemy==1.4.44
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.1.1
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.13.0
torchmetrics==0.11.0
torchsummary==1.5.1
torchviz==0.0.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.7
Werkzeug==2.2.2
x-transformers==1.6.3
yarl==1.8.2
zipp==3.11.0
